Refactor Fuqu types, add template args (breaking!)
D - is the data argument (was always any, now user defined)
O - is the original message (pubsub Message for pubsub, else any)

User can now omit the type arguments (leave default) and work with any
as they used to now, allowing for easy handler dynamic assignemnt.

The usual case is that each class will have a fixed message type. User
can now define that type in argument and work with class in a typesafe
manner.

BREAKING UPDATE: It was necessary before the update to wrap messages in
`{data: ...}`. Now this depends on the type, so implicit onwraping had
been removed.